refactor(server): job core into dedicated service calls
