+++ server/apps/immich/src/modules/schedule-tasks/schedule-tasks.service.ts
@@ -1,12 +1,13 @@
-import { JobName, JobService } from '@app/domain';
+import { UserService } from '@app/domain';
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 
 @Injectable()
 export class ScheduleTasksService {
-  constructor(private jobService: JobService) {}
+  constructor(private userService: UserService) {}
+
   @Cron(CronExpression.EVERY_DAY_AT_11PM)
   async onUserDeleteCheck() {
-    await this.jobService.handle({ name: JobName.USER_DELETE_CHECK });
+    await this.userService.handleUserDeleteCheck();
   }
 }
