+++ server/libs/domain/src/asset/index.ts
@@ -1,2 +1,3 @@
 export * from './asset.repository';
+export * from './asset.service';
 export * from './response-dto';
