+++ server/libs/domain/src/domain.module.ts
@@ -1,19 +1,23 @@
 import { DynamicModule, Global, Module, ModuleMetadata, Provider } from '@nestjs/common';
 import { APIKeyService } from './api-key';
+import { AssetService } from './asset';
 import { AuthService } from './auth';
 import { DeviceInfoService } from './device-info';
-import { JobService } from './job';
 import { OAuthService } from './oauth';
 import { ShareService } from './share';
+import { StorageService } from './storage';
+import { StorageTemplateService } from './storage-template';
 import { INITIAL_SYSTEM_CONFIG, SystemConfigService } from './system-config';
 import { UserService } from './user';
 
 const providers: Provider[] = [
+  AssetService,
   APIKeyService,
   AuthService,
   DeviceInfoService,
-  JobService,
   OAuthService,
+  StorageService,
+  StorageTemplateService,
   SystemConfigService,
   UserService,
   ShareService,
