+++ server/libs/domain/src/job/index.ts
@@ -1,4 +1,3 @@
 export * from './job.constants';
 export * from './job.interface';
 export * from './job.repository';
-export * from './job.service';
