+++ server/libs/domain/src/storage-template/index.ts
@@ -1,0 +1,2 @@
+export * from './storage-template.core';
+export * from './storage-template.service';
