+++ server/libs/domain/src/storage/index.ts
@@ -1,2 +1,2 @@
-export * from './storage.core';
 export * from './storage.repository';
+export * from './storage.service';
