+++ server/libs/infra/src/infra.module.ts
@@ -84,16 +84,7 @@ const providers: Provider[] = [
         },
       }),
     }),
-    BullModule.registerQueue(
-      { name: QueueName.USER_DELETION },
-      { name: QueueName.THUMBNAIL_GENERATION },
-      { name: QueueName.ASSET_UPLOADED },
-      { name: QueueName.METADATA_EXTRACTION },
-      { name: QueueName.VIDEO_CONVERSION },
-      { name: QueueName.MACHINE_LEARNING },
-      { name: QueueName.CONFIG },
-      { name: QueueName.BACKGROUND_TASK },
-    ),
+    BullModule.registerQueue(...Object.values(QueueName).map((name) => ({ name }))),
   ],
   providers: [...providers],
   exports: [...providers, BullModule],
