+++ server/libs/infra/src/job/job.repository.ts
@@ -1,14 +1,4 @@
-import {
-  IAssetJob,
-  IAssetUploadedJob,
-  IJobRepository,
-  IMetadataExtractionJob,
-  IUserDeletionJob,
-  JobCounts,
-  JobItem,
-  JobName,
-  QueueName,
-} from '@app/domain';
+import { IAssetJob, IJobRepository, IMetadataExtractionJob, JobCounts, JobItem, JobName, QueueName } from '@app/domain';
 import { InjectQueue } from '@nestjs/bull';
 import { BadRequestException, Logger } from '@nestjs/common';
 import { Queue } from 'bull';
@@ -17,13 +7,11 @@ export class JobRepository implements IJobRepository {
   private logger = new Logger(JobRepository.name);
 
   constructor(
-    @InjectQueue(QueueName.ASSET_UPLOADED) private assetUploaded: Queue<IAssetUploadedJob>,
     @InjectQueue(QueueName.BACKGROUND_TASK) private backgroundTask: Queue,
     @InjectQueue(QueueName.MACHINE_LEARNING) private machineLearning: Queue<IAssetJob>,
     @InjectQueue(QueueName.METADATA_EXTRACTION) private metadataExtraction: Queue<IMetadataExtractionJob>,
-    @InjectQueue(QueueName.CONFIG) private storageMigration: Queue,
+    @InjectQueue(QueueName.STORAGE_TEMPLATE_MIGRATION) private storageTemplateMigration: Queue,
     @InjectQueue(QueueName.THUMBNAIL_GENERATION) private thumbnail: Queue,
-    @InjectQueue(QueueName.USER_DELETION) private userDeletion: Queue<IUserDeletionJob>,
     @InjectQueue(QueueName.VIDEO_CONVERSION) private videoTranscode: Queue<IAssetJob>,
   ) {}
 
@@ -43,7 +31,7 @@ export class JobRepository implements IJobRepository {
   async queue(item: JobItem): Promise<void> {
     switch (item.name) {
       case JobName.ASSET_UPLOADED:
-        await this.assetUploaded.add(item.name, item.data, { jobId: item.data.asset.id });
+        await this.backgroundTask.add(item.name, item.data, { jobId: item.data.asset.id });
         break;
 
       case JobName.DELETE_FILES:
@@ -61,18 +49,21 @@ export class JobRepository implements IJobRepository {
         await this.metadataExtraction.add(item.name, item.data);
         break;
 
-      case JobName.TEMPLATE_MIGRATION:
-      case JobName.CONFIG_CHANGE:
-        await this.storageMigration.add(item.name, {});
-        break;
-
       case JobName.GENERATE_JPEG_THUMBNAIL:
       case JobName.GENERATE_WEBP_THUMBNAIL:
         await this.thumbnail.add(item.name, item.data);
         break;
 
       case JobName.USER_DELETION:
-        await this.userDeletion.add(item.name, item.data);
+        await this.backgroundTask.add(item.name, item.data);
+        break;
+
+      case JobName.STORAGE_TEMPLATE_MIGRATION:
+        await this.storageTemplateMigration.add(item.name);
+        break;
+
+      case JobName.SYSTEM_CONFIG_CHANGE:
+        await this.backgroundTask.add(item.name, {});
         break;
 
       case JobName.VIDEO_CONVERSION:
@@ -87,14 +78,14 @@ export class JobRepository implements IJobRepository {
 
   private getQueue(name: QueueName) {
     switch (name) {
+      case QueueName.STORAGE_TEMPLATE_MIGRATION:
+        return this.storageTemplateMigration;
       case QueueName.THUMBNAIL_GENERATION:
         return this.thumbnail;
       case QueueName.METADATA_EXTRACTION:
         return this.metadataExtraction;
       case QueueName.VIDEO_CONVERSION:
         return this.videoTranscode;
-      case QueueName.CONFIG:
-        return this.storageMigration;
       case QueueName.MACHINE_LEARNING:
         return this.machineLearning;
       default:
